Replace AudioConferenceMixer with AudioMixer.
This CL re-routes audio through AudioMixer instead of AudioConferenceMixer.
This is done without any modifications to VoiceEngine.

Previously, output audio was polled by an AudioDevice through an AudioTransport
pointer, which was an instance of VoEBaseImpl. VoiceEngineImpl sent the
request for data on to OutputMixer and further to AudioConferenceMixer.

This CL changes the audio flow to an AudioDevice. We reconfigure the AudioDevice
to have another AudioTransport pointer, which points to an AudioTransportProxy.
The AudioTransportProxy is responsible for feeding mixed data to the
AudioProcessing component for echo cancellation, and to resample the audio data
after AudioProcessing and before it is sent to the AudioDevice.

The set up of the audio path was previously done during VoiceEngine
initialization. Now it is changed in the AudioState constructor.

This list shows where audio-path-related VoiceEngine functionality has been
moved:

OutputMixer --> AudioTransportProxy
VoiceEngineImpl --> AudioState, AudioTransportProxy
SharedData --> AudioState
Channel --> AudioReceiveStream, ChannelProxy, Channel

AudioState owns the new mixer and connects it to AudioTransport and
AudioDevice on initialization.

The audio input source is AudioReceiveStream, which  registers itself with the
mixer (which it gets from AudioState) on Start and Stop.

# Since the AudioTransport interface contains non-const references.
NOPRESUBMIT=True
BUG=webrtc:6346

Review-Url: https://codereview.webrtc.org/2436033002
Cr-Original-Commit-Position: refs/heads/master@{#15193}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 04c0722cf1fc38e6888ef44e413603e0cb461ea4